Revert "[TACHYON-108] Remote write support"
