+++ clients/unshaded/src/main/java/tachyon/client/RemoteBlockInStream.java
@@ -207,7 +207,7 @@ public class RemoteBlockInStream extends BlockInStream {
     // attempts that are invalidated later due to seek/skips
     if (bytesLeft > 0 && mBlockOutStream == null && mRecache) {
       try {
-        mBlockOutStream = BlockOutStream.get(mFile, WriteType.TRY_CACHE, mBlockIndex, mTachyonConf);
+        mBlockOutStream = new BlockOutStream(mFile, WriteType.TRY_CACHE, mBlockIndex, mTachyonConf);
       } catch (IOException ioe) {
         LOG.warn("Recache attempt failed.", ioe);
         cancelRecache();
