+++ common/src/main/java/tachyon/Constants.java
@@ -211,7 +211,6 @@ public class Constants {
       "tachyon.user.remote.read.buffer.size.byte";
   public static final String USER_DEFAULT_WRITE_TYPE = "tachyon.user.file.writetype.default";
   public static final String USER_REMOTE_BLOCK_READER = "tachyon.user.remote.block.reader.class";
-  public static final String USER_REMOTE_BLOCK_WRITER = "tachyon.user.remote.block.writer.class";
   public static final String USER_ENABLE_LOCAL_READ = "tachyon.user.localread.enable";
   public static final String USER_ENABLE_LOCAL_WRITE = "tachyon.user.localwrite.enable";
