+++ common/src/main/java/tachyon/network/protocol/RPCBlockRequest.java
@@ -32,7 +32,6 @@ public class RPCBlockRequest extends RPCRequest {
   private final long mOffset;
   private final long mLength;
 
-  // TODO: rename this to RPCBlockReadRequest.
   public RPCBlockRequest(long blockId, long offset, long length) {
     mBlockId = blockId;
     mOffset = offset;
@@ -61,7 +60,7 @@ public class RPCBlockRequest extends RPCRequest {
   @Override
   public int getEncodedLength() {
     // TODO: adjust the length when client also uses netty.
-    // 3 longs (mBlockId, mOffset, mLength) + 1 short (DATA_SERVER_REQUEST_MESSAGE)
+    // 3 longs (mBLockId, mOffset, mLength) + 1 short (DATA_SERVER_REQUEST_MESSAGE)
     return Longs.BYTES * 3 + Shorts.BYTES;
   }
