+++ common/src/main/resources/tachyon-default.properties
@@ -84,6 +84,5 @@ tachyon.user.default.block.size.byte=512MB
 tachyon.user.quota.unit.bytes=8MB
 tachyon.user.file.buffer.bytes=1MB
 tachyon.user.remote.block.reader.class=tachyon.client.tcp.TCPRemoteBlockReader
-tachyon.user.remote.block.writer.class=tachyon.client.netty.NettyRemoteBlockWriter
 tachyon.user.remote.read.buffer.size.byte=8MB
 tachyon.user.network.netty.worker.threads=0
