+++ servers/src/main/java/tachyon/worker/netty/DataServerHandler.java
@@ -16,9 +16,7 @@
 package tachyon.worker.netty;
 
 import java.io.IOException;
-import java.io.RandomAccessFile;
 import java.nio.ByteBuffer;
-import java.nio.MappedByteBuffer;
 import java.nio.channels.FileChannel;
 
 import com.google.common.base.Optional;
@@ -34,24 +32,18 @@ import io.netty.channel.ChannelHandlerContext;
 import io.netty.channel.SimpleChannelInboundHandler;
 
 import tachyon.Constants;
-import tachyon.StorageLevelAlias;
 import tachyon.Users;
 import tachyon.conf.TachyonConf;
 import tachyon.network.protocol.RPCBlockRequest;
 import tachyon.network.protocol.RPCBlockResponse;
-import tachyon.network.protocol.RPCBlockWriteRequest;
-import tachyon.network.protocol.RPCBlockWriteResponse;
 import tachyon.network.protocol.RPCMessage;
 import tachyon.network.protocol.RPCRequest;
 import tachyon.network.protocol.RPCResponse;
 import tachyon.network.protocol.databuffer.DataBuffer;
 import tachyon.network.protocol.databuffer.DataByteBuffer;
 import tachyon.network.protocol.databuffer.DataFileChannel;
-import tachyon.util.CommonUtils;
 import tachyon.worker.block.BlockDataManager;
 import tachyon.worker.block.io.BlockReader;
-import tachyon.worker.block.io.BlockWriter;
-import tachyon.worker.block.meta.BlockMetaBase;
 
 /**
  * This class has the main logic of the read path to process {@link RPCRequest} messages and return
@@ -78,9 +70,6 @@ public final class DataServerHandler extends SimpleChannelInboundHandler<RPCMess
     switch (msg.getType()) {
       case RPC_BLOCK_REQUEST:
         handleBlockRequest(ctx, (RPCBlockRequest) msg);
-        break;
-      case RPC_BLOCK_WRITE_REQUEST:
-        handleBlockWriteRequest(ctx, (RPCBlockWriteRequest) msg);
         break;
       default:
         throw new IllegalArgumentException("No handler implementation for rpc msg type: "
@@ -134,53 +123,6 @@ public final class DataServerHandler extends SimpleChannelInboundHandler<RPCMess
       }
     } finally {
       mDataManager.unlockBlock(lockId);
-    }
-  }
-
-  // TODO: This write request handler is very simple in order to be stateless. Therefore, the block
-  // file is opened and closed for every request. If this is too slow, then this handler should be
-  // optimized to keep state.
-  private void handleBlockWriteRequest(final ChannelHandlerContext ctx,
-      final RPCBlockWriteRequest req) throws IOException {
-    final long userId = req.getUserId();
-    final long blockId = req.getBlockId();
-    final long offset = req.getOffset();
-    final long length = req.getLength();
-    final DataBuffer data = req.getPayloadDataBuffer();
-
-    BlockWriter writer = null;
-    try {
-      req.validate();
-      ByteBuffer buffer = data.getReadOnlyByteBuffer();
-      if (buffer.remaining() <= 0) {
-        throw new IOException("Empty buffer to write.");
-      }
-
-      if (offset == 0) {
-        // This is the first write to the block, so create the temp block file. The file will only
-        // be created if the first write starts at offset 0. This allocates enough space for the
-        // write.
-        mDataManager.createBlockRemote(userId, blockId, StorageLevelAlias.MEM.getValue(), length);
-      } else {
-        // Allocate enough space in the existing temporary block for the write.
-        mDataManager.requestSpace(userId, blockId, length);
-      }
-      writer = mDataManager.getTempBlockWriterRemote(userId, blockId);
-      writer.append(buffer);
-
-      ChannelFuture future =
-          ctx.writeAndFlush(new RPCBlockWriteResponse(userId, blockId, offset, length, true));
-      future.addListener(ChannelFutureListener.CLOSE);
-      future.addListener(new ClosableResourceChannelListener(writer));
-    } catch (Exception e) {
-      LOG.error("Error writing remote block : " + e.getMessage(), e);
-      RPCBlockWriteResponse resp =
-          new RPCBlockWriteResponse(userId, blockId, offset, length, false);
-      ChannelFuture future = ctx.writeAndFlush(resp);
-      future.addListener(ChannelFutureListener.CLOSE);
-      if (writer != null) {
-        writer.close();
-      }
     }
   }
