+++ servers/src/main/java/tachyon/worker/netty/PipelineHandler.java
@@ -20,18 +20,22 @@ import io.netty.channel.ChannelPipeline;
 import io.netty.channel.socket.SocketChannel;
 import io.netty.handler.stream.ChunkedWriteHandler;
 
+import tachyon.conf.TachyonConf;
 import tachyon.network.protocol.RPCMessage;
 import tachyon.network.protocol.RPCMessageDecoder;
 import tachyon.network.protocol.RPCMessageEncoder;
+import tachyon.worker.block.BlockDataManager;
 
 /**
  * Adds the block server's pipeline into the channel.
  */
 public final class PipelineHandler extends ChannelInitializer<SocketChannel> {
-  private final DataServerHandler mDataServerHandler;
+  private final BlockDataManager mDataManager;
+  private final TachyonConf mTachyonConf;
 
-  public PipelineHandler(final DataServerHandler handler) {
-    mDataServerHandler = handler;
+  public PipelineHandler(final BlockDataManager dataManager, final TachyonConf tachyonConf) {
+    mDataManager = dataManager;
+    mTachyonConf = tachyonConf;
   }
 
   @Override
@@ -41,6 +45,6 @@ public final class PipelineHandler extends ChannelInitializer<SocketChannel> {
     pipeline.addLast("frameDecoder", RPCMessage.createFrameDecoder());
     pipeline.addLast("RPCMessageDecoder", new RPCMessageDecoder());
     pipeline.addLast("RPCMessageEncoder", new RPCMessageEncoder());
-    pipeline.addLast("dataServerHandler", mDataServerHandler);
+    pipeline.addLast("dataServerHandler", new DataServerHandler(mDataManager, mTachyonConf));
   }
 }
